Add isTimeViewSticky attriubte and change progress slider behaviour on touch
isTimeViewSticky attribute when set to true will stick the time views to the start and end of the range slider, otherwise it will track with the start and end indicator views.

ProgressIndicator will now hide when the start/end indicators are being dragged and appear on drag ended.
The startTimeView time will also reflect the progress indicators value.